% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rfcvquery.R
\name{get.intro.name}
\alias{get.intro.name}
\title{Return introduction names.}
\usage{
get.intro.name(conn, intro.codes)
}
\arguments{
\item{conn}{a DBIConnection object, as produced by dbConnect}

\item{intro.codes}{vector of introduction codes ("CodeIntro")}
}
\value{
vector of introduction names
}
\description{
http://bioweb.ensam.inra.fr/collections_vigne/Intro.php
}
\author{
Timothée Flutre
}

