% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rfcvquery.R
\name{get.variety.code.name}
\alias{get.variety.code.name}
\title{Return variety codes and names.}
\usage{
get.variety.code.name(conn, intro.codes)
}
\arguments{
\item{conn}{a DBIConnection object, as produced by dbConnect}

\item{intro.codes}{vector of introduction codes ("CodeIntro")}
}
\value{
vector of variety names
}
\description{
http://bioweb.ensam.inra.fr/collections_vigne/Variete.php
}
\author{
Timothée Flutre
}

