% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfcvquery.R
\name{getIntroNames}
\alias{getIntroNames}
\title{Introduction/accession level}
\usage{
getIntroNames(conn, intro.codes)
}
\arguments{
\item{conn}{a DBIConnection object, as produced by dbConnect}

\item{intro.codes}{vector of introduction codes ("CodeIntro")}
}
\value{
data frame of introduction codes and names
}
\description{
Return the introduction/accession names from a set of introduction/accession codes.
}
\examples{
\dontrun{## obviously, you must have read access to the database
library(rfcvquery)
library(getPass)
conn <- dbConnect(drv=MySQL(), host="...", dbname="...",
                  user="...", password=getPass())
getIntroNames(conn, c("23297Mtp13", "23298Mtp28"))
on.exit(dbDisconnect(conn))
}
}
\author{
Timothée Flutre
}
