% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfcvquery.R
\name{getTemplateInsertVariety}
\alias{getTemplateInsertVariety}
\title{Variety/Population level}
\usage{
getTemplateInsertVariety(conn)
}
\arguments{
\item{conn}{a DBIConnection object, as produced by dbConnect}
}
\value{
data.frame
}
\description{
Return a template of what to provide to insert a new variety.
}
\author{
Timothee Flutre
}
