% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfcvquery.R
\name{getVarietyCodeName}
\alias{getVarietyCodeName}
\title{Variety/Population level}
\usage{
getVarietyCodeName(conn, intro.codes)
}
\arguments{
\item{conn}{a DBIConnection object, as produced by dbConnect}

\item{intro.codes}{vector of introduction codes ("CodeIntro")}
}
\value{
data.frame of variety codes and names
}
\description{
Return the variety/population codes and names from a set of introduction/accession codes.
}
\author{
Timothée Flutre
}

