% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfcvquery.R
\name{getVarietyCodesNames}
\alias{getVarietyCodesNames}
\title{Variety/Population level}
\usage{
getVarietyCodesNames(conn, intro.codes)
}
\arguments{
\item{conn}{a DBIConnection object, as produced by dbConnect}

\item{intro.codes}{vector of introduction codes ("CodeIntro")}
}
\value{
data frame of variety code(s) and name(s), with one row per input intro code (even if there are duplicates)
}
\description{
Return the variety/population codes and names from a set of introduction/accession codes.
}
\examples{
\dontrun{## obviously, you must have read access to the database
library(rfcvquery)
library(getPass)
conn <- dbConnect(drv=MySQL(), host="...", dbname="...",
                  user="...", password=getPass())
getVarietyCodeName(conn, c("23297Mtp13", "23298Mtp28"))
on.exit(dbDisconnect(conn))
}
}
\author{
Timothée Flutre
}
