% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfcvquery.R
\name{getVarietyColors}
\alias{getVarietyColors}
\title{Skin and pulp colors}
\usage{
getVarietyColors(conn, variety.codes)
}
\arguments{
\item{conn}{a DBIConnection object, as produced by dbConnect}

\item{variety.codes}{vector of variety codes ("CodeVar")}
}
\value{
data.frame of variety codes, names and colors
}
\description{
Return the skin and pul colors from a set of variety codes.
}
\examples{
\dontrun{## obviously, you must have read access to the database
library(RMySQL)
library(getPass)
conn <- dbConnect(drv=MySQL(), host="...", dbname="...",
                  user="...", password=getPass())
variety.codes <- c("18", "300", "195")
getVarietyColor(conn, variety.codes)
on.exit(dbDisconnect(con))
}
}
\author{
Timothee Flutre
}
