% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{package}
\name{rfcvquery}
\alias{rfcvquery}
\alias{rfcvquery-package}
\title{rfcvquery: A package to query the RFCV database.}
\description{
To use this package, a "read" access to the MySQL database is required.
}
\section{Encoding}{

Note that the content of the database is in the \href{https://en.wikipedia.org/wiki/ISO/IEC_8859}{ISO-8859-1} standard of character encoding.
As such, once the result of a query is available, it may be useful to use the following function: \code{res.new <- iconv(res, from="ISO-8859-1", to="UTF-8")}.
}

