% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfcvquery.R
\name{getIntroName}
\alias{getIntroName}
\title{Introduction/accession level}
\usage{
getIntroName(conn, intro.codes)
}
\arguments{
\item{conn}{a DBIConnection object, as produced by dbConnect}

\item{intro.codes}{vector of introduction codes ("CodeIntro")}
}
\value{
vector of introduction names
}
\description{
Return the introduction/accession names from a set of introduction/accession codes.
}
\author{
Timothée Flutre
}

