% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfcvquery.R
\name{getVarietyCodeName}
\alias{getVarietyCodeName}
\title{Variety/Population level}
\usage{
getVarietyCodeName(conn, intro.codes)
}
\arguments{
\item{conn}{a DBIConnection object, as produced by dbConnect}

\item{intro.codes}{vector of introduction codes ("CodeIntro")}
}
\value{
data.frame of variety codes and names
}
\description{
Return the variety/population codes and names from a set of introduction/accession codes.
}
\examples{
\dontrun{## obviously, you must have read access to the database
library(RMySQL)
library(getPass)
conn <- dbConnect(drv=MySQL(), host="...", dbname="...",
                  user="...", password=getPass())
getVarietyCodeName(conn, c("23297Mtp13", "23298Mtp28"))
on.exit(dbDisconnect(con))
}
}
\author{
Timothée Flutre
}
